+++ examples/01_encodings.py
@@ -267,8 +267,7 @@ print(f"mean fit time: {np.mean(results['fit_time']):.3f} seconds")
 from skrub import MinHashEncoder, ToCategorical
 
 vectorizer = TableVectorizer(
-    low_cardinality_transformer=ToCategorical(),
-    high_cardinality_transformer=MinHashEncoder(),
+    low_cardinality=ToCategorical(), high_cardinality=MinHashEncoder()
 )
 pipeline = make_pipeline(
     vectorizer, HistGradientBoostingRegressor(categorical_features="from_dtype")