+++ skrub/tests/test_table_vectorizer.py
@@ -248,10 +248,10 @@ X_tuples = [
 
 def passthrough_vectorizer():
     return TableVectorizer(
-        high_cardinality_transformer="passthrough",
-        low_cardinality_transformer="passthrough",
-        numeric_transformer="passthrough",
-        datetime_transformer="passthrough",
+        high_cardinality="passthrough",
+        low_cardinality="passthrough",
+        numeric="passthrough",
+        datetime="passthrough",
     )
 
 
@@ -374,7 +374,7 @@ inputs = [
 
 def test_handle_unknown_category():
     X = _get_clean_dataframe()
-    # Treat all columns as low cardinality
+    # Treat all columns as having few unique values
     table_vec = TableVectorizer(cardinality_threshold=7).fit(X)
     X_unknown = pd.DataFrame(
         {
@@ -423,7 +423,7 @@ def test_handle_unknown_category():
     [
         TableVectorizer(),
         TableVectorizer(
-            low_cardinality_transformer=MinHashEncoder(),
+            low_cardinality=MinHashEncoder(),
         ),
     ],
 )
@@ -506,7 +506,7 @@ def test_changing_types(X_train, X_test, expected_X_out):
     """
     table_vec = TableVectorizer(
         # only extract the total seconds
-        datetime_transformer=DatetimeEncoder(resolution=None)
+        datetime=DatetimeEncoder(resolution=None)
     )
     table_vec.fit(X_train)
     X_out = table_vec.transform(X_test)
@@ -536,7 +536,7 @@ def test_column_by_column():
         pytest.xfail("pandas is_string_dtype incorrect in old pandas")
     X = _get_clean_dataframe()
     vectorizer = TableVectorizer(
-        high_cardinality_transformer=GapEncoder(n_components=2, random_state=0),
+        high_cardinality=GapEncoder(n_components=2, random_state=0),
         cardinality_threshold=4,
     )
     X_trans = vectorizer.fit_transform(X)
@@ -551,7 +551,7 @@ def test_column_by_column():
 
 @skip_if_no_parallel
 @pytest.mark.parametrize(
-    "high_cardinality_transformer",
+    "high_cardinality",
     # The GapEncoder and the MinHashEncoder should be parallelized on all columns.
     # The OneHotEncoder should not be parallelized.
     [
@@ -560,10 +560,10 @@ def test_column_by_column():
         MinHashEncoder(n_components=2),
     ],
 )
-def test_parallelism(high_cardinality_transformer):
+def test_parallelism(high_cardinality):
     X = _get_clean_dataframe()
     params = dict(
-        high_cardinality_transformer=high_cardinality_transformer,
+        high_cardinality=high_cardinality,
         cardinality_threshold=4,
     )
     vectorizer = TableVectorizer(**params)
@@ -607,9 +607,9 @@ def test_pandas_sparse_array():
 def test_wrong_transformer():
     X = _get_clean_dataframe()
     with pytest.raises(ValueError):
-        TableVectorizer(high_cardinality_transformer="passthroughtypo").fit(X)
+        TableVectorizer(high_cardinality="passthroughtypo").fit(X)
     with pytest.raises(TypeError):
-        TableVectorizer(high_cardinality_transformer=None).fit(X)
+        TableVectorizer(high_cardinality=None).fit(X)
 
 
 invalid_tuples = [
@@ -678,7 +678,7 @@ def test_accept_pipeline():
     # non-regression test for https://github.com/skrub-data/skrub/issues/886
     # TableVectorizer used to force transformers to inherit from TransformerMixin
     df = pd.DataFrame(dict(a=[1.1, 2.2]))
-    tv = TableVectorizer(numeric_transformer=make_pipeline("passthrough"))
+    tv = TableVectorizer(numeric=make_pipeline("passthrough"))
     tv.fit(df)
 
 
@@ -720,5 +720,5 @@ def test_supervised_encoder(df_module):
     # of the defaults encoders do)
     X = df_module.make_dataframe({"a": [f"c_{i}" for _ in range(5) for i in range(4)]})
     y = np.random.default_rng(0).normal(size=sbd.shape(X)[0])
-    tv = TableVectorizer(low_cardinality_transformer=TargetEncoder())
+    tv = TableVectorizer(low_cardinality=TargetEncoder())
     tv.fit_transform(X, y)